Spawn target objects and blocker added
After the countdown, the blocker and target object is initialized. With every click to the right hand primary controller (if the enum setting is set to distance perception) the distance between the endpoint of the arc and object is calculated. The object is destroyed and another object is instantiated at the next coordinate.

TODO: Process needs to be streamlined, spawn the object, click ready (followed by the 3 second countdown), block view