Distance perception metrics output to csv
The values are outputted to a csv file through CSV Writer and DataManager.

TODO: Add a Technique Manager to determine the active technique. A system needs to be implemented to determine which interaction technique is going to be activated and record to the csv (which currently has the value hard-coded)