Target Object Task Complete Callback
Dense batch target object contains a script to trigger callback when the object has been in collision with the target surface for 1 second. This is going to be the task definition and the object instantiation cycle will continue.

TODO: Finalize the object creation cycle for dense batch step
TODO: Make sure everything else works